requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2024.6.2=h56e8100_0
certifi=2021.5.30=py36haa95532_0
freeglut=3.2.2=h0e60522_1
freetype=2.12.1=h546665d_0
giflib=5.2.1=h8cc25b3_3
icu=69.1=h0e60522_0
intel-openmp=2024.1.0=h57928b3_966
jasper=2.0.33=h77af90b_0
jpeg=9e=h8ffe710_2
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=22_win64_mkl
libcblas=3.9.0=22_win64_mkl
libclang=13.0.1=default_h81446c8_0
libdeflate=1.10=h8ffe710_0
liblapack=3.9.0=22_win64_mkl
liblapacke=3.9.0=22_win64_mkl
libopencv=4.5.3=py36h530cdb0_2
libpng=1.6.37=h1d00b33_4
libprotobuf=3.16.0=h7755175_0
libtiff=4.3.0=hc4061b1_4
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h8ffe710_0
libzlib=1.2.12=h8ffe710_2
mkl=2024.1.0=h66d3029_692
numpy=1.19.5=py36h4b40d73_2
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.3=py36ha15d459_2
openjpeg=2.5.0=hb211442_0
openssl=1.1.1w=h2bbff1b_0
pillow=8.3.2=py36h10c25d6_0
pip=21.2.2=py36haa95532_0
py-opencv=4.5.3=py36h5098b23_2
python=3.6.13=h3758d61_0
python_abi=3.6=2_cp36m
qt=5.12.9=h556501e_6
setuptools=58.0.4=py36haa95532_0
sqlite=3.45.3=h2bbff1b_0
tbb=2021.5.0=h2d74725_1
tk=8.6.12=h2bbff1b_0
ucrt=10.0.22621.0=h57928b3_0
vc=14.2=h2eaa2aa_4
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.6=h8d14728_0
zlib=1.2.12=h8ffe710_2
zstd=1.5.2=h7755175_4
